requirements.txt
altair                5.0.1
asgiref               3.7.2
attrs                 23.1.0
beautifulsoup4        4.12.2
blinker               1.6.2
bs4                   0.0.1
cachetools            5.3.1
certifi               2023.5.7
charset-normalizer    3.1.0
click                 8.1.3
colorama              0.4.6
contourpy             1.1.0
cycler                0.11.0
decorator             5.1.1
Django                4.2.2
djangorestframework   3.14.0
fonttools             4.40.0
gitdb                 4.0.10
GitPython             3.1.31
idna                  3.4
importlib-metadata    6.7.0
Jinja2                3.1.2
jsonschema            4.17.3
kiwisolver            1.4.4
markdown-it-py        3.0.0
MarkupSafe            2.1.3
matplotlib            3.7.1
mdurl                 0.1.2
numpy                 1.25.0
packaging             23.1
pandas                2.0.2
Pillow                9.5.0
pip                   21.2.3
protobuf              4.23.3
psycopg2              2.9.6
pyarrow               12.0.1
pydeck                0.8.1b0
Pygments              2.15.1
Pympler               1.0.1
pyparsing             3.1.0
pyrsistent            0.19.3
python-dateutil       2.8.2
pytz                  2023.3
pytz-deprecation-shim 0.1.0.post0
requests              2.31.0
rich                  13.4.2
setuptools            57.4.0
six                   1.16.0
smmap                 5.0.0
soupsieve             2.4.1